Move code related to MultiTxId to separate files
